Computer-Rename.ps1
When you run the script it will check if you have ran the script all the way previously based upon the site prefix. If you have not it will help you build a computer name. Once the computer name is built it will check if it exists inside of AD and if it does not it will rename the computer to the new name.
    When you run the script after the first time time it runs a quick check to see if the prefix is the same of what you selected before. If it is the script checks if the computer is joined to a domain, if so it exits. If not it will jooin it to the domain and reboot.